requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-ckeditor==6.1.0
django-countries==7.3.2
django-js-asset==1.2.2
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
django-xhtml2pdf==0.0.4
future==0.18.2
html5lib==1.1
idna==2.10
Pillow==8.2.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
pyuca==1.2
reportlab==3.6.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
typing_extensions==4.1.1
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2.5
